requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.2.0
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
setuptools==69.0.3
sqlparse==0.5.5
uritemplate==4.2.0
wheel==0.42.0
whitenoise==6.11.0
